Add other types of attachment
There are some other types of attachment, which doesn't correspond with types of statuses
and they are not documented, because on FB Graph API documentation they state that
"attachment" field is deprecated and not used anymore, but it isn't and is still used.
There is no list of available attachment types, so some types could be unrecognised.
In that case they are labeled as "status" which means there is no attachment showed.