Separate all author info from data to users table
I moved info about authors from data table (from_id and from_name) to users table,
because most of the authors would be also users. It is better for three reasons:
1) Author data is separated from post - makes logically better sense
2) Author data won't be duplicated
3) Author will be connected to exact user - better for handling user info etc.

I've created new column "registered" in user's table.
When it is 1, user is registered and can have access
to fitak and has his own user profile. When it is 0, it's only author which is crawled from FB,
but when this person create account and connect it with FB, it will match his author profile
(not implemented yet).

I also changed corresponding things in ORM, FB API, FB crawler, stream template, elastic
so everything works now.