Connect user profile with FB account when registering
There needed to be possibility to connect FB account with internal user account in Fitak
for ability to make posts which would be send back to FB group and possibility to
comment/reply to comments on posts which are crawled from FB.

FB login doesn't supply internal user account (so it's not the another way to log into Fitak),
but it retrieves FB id, user access token, FB name and profile picture and add that informations
into register form.

When there is already some non-registered user in DB (it means that somebody written into FB group
which was crawled by Fitak afterwards, but he doesn't have Fitak account yet) and if he connect his account
with FB, it will match corresponding user with the same ID and just fill more informations to this profile.

There is some other bug fixes (as catching exceptions when storing duplicate entry to DB or when sending
 e-mail doesn't work properly).